// src/features/auth/schema.ts
import { z } from "zod"

export const signupSchema = z.object({
  name: z.string().min(3,"Name must be at least 3 characters"),
  email: z.string().email("Invalid email"),
  password: z.string().min(6,"Password must be at least 6 characters"),
  role: z.enum(["customer","seller","admin"])
})

export const loginSchema = z.object({
  email: z.string().email("Invalid email"),
  password: z.string().min(6,"Password must be at least 6 characters")
})